% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_clean.R
\name{aggregate_doc_clean_data}
\alias{aggregate_doc_clean_data}
\title{Title}
\usage{
aggregate_doc_clean_data(.doc_data, .doc_aggregate_fn)
}
\arguments{
\item{.doc_data}{...}

\item{.doc_aggregate_fn}{...}
}
\value{
...
}
\description{
Title
}
\examples{
if(FALSE){
  NULL
}
}
