% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_index.R
\name{gather_doc_index_data}
\alias{gather_doc_index_data}
\title{Title}
\usage{
gather_doc_index_data(
  .years,
  .quarters = 1:4,
  .form_types = "10-K",
  .data_dir = NULL
)
}
\arguments{
\item{.years}{...}

\item{.quarters}{...}

\item{.form_types}{...}

\item{.data_dir}{...
#'}
}
\value{
A tibble ...
}
\description{
Title
}
\examples{
if(FALSE){
  doc_index <- get_doc_index(.years=2021:2022)
}
}
