% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_clean.R
\name{gather_doc_stat_body_data}
\alias{gather_doc_stat_body_data}
\title{Title}
\usage{
gather_doc_stat_body_data(.doc_data, .doc_stat_body_fn, .n_workers = NULL)
}
\arguments{
\item{.doc_data}{...}

\item{.doc_stat_body_fn}{...}

\item{.n_workers}{...}
}
\value{
...
}
\description{
Title
}
\examples{
NULL
}
