% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_dir}
\alias{get_data_dir}
\title{Title}
\usage{
get_data_dir(.doc_data)
}
\arguments{
\item{.doc_data}{...}
}
\value{
data_dir
}
\description{
Title
}
\examples{
if(FALSE){
  data_dir <- get_data_dir(doc_data)
}
}
