% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_storage_rel_path}
\alias{get_storage_rel_path}
\title{Title ...}
\usage{
get_storage_rel_path(.file_set, .com_cik, .doc_an)
}
\arguments{
\item{.file_set}{character ....}

\item{.com_cik}{character ...}

\item{.doc_an}{character ...}
}
\value{
...
}
\description{
Title ...
}
