% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_data_dir}
\alias{set_data_dir}
\title{Title}
\usage{
set_data_dir(.doc_data, .data_dir)
}
\arguments{
\item{.doc_data}{...}

\item{.data_dir}{...}
}
\value{
...
}
\description{
Title
}
\section{Notes}{

LALLALALAL
}

\examples{
NULL
}
