% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_retries}
\alias{with_retries}
\title{Title}
\usage{
with_retries(.function, .return_error = FALSE)
}
\arguments{
\item{.function}{...}

\item{.return_error}{...}
}
\value{
A copy of .function
}
\description{
MODIFY A FUNCTION SO THAT IT REEXECUTES IF AN ERROR IS THROWN THAT CONTAINS
"TIMEOUT WAS REACHED" OR "COULD NOT RESOLVE HOST". IF AN OTHER ERROR IS
THROWN RETURN THAT ERROR OR STOP (depending on .return_error)
}
\examples{
NULL
}
